Refactor new examples
